# Fix the latency issue caused after not using anonymous user while reading or downloading image from registry